chore(generate_prototypes_from_TODO.yml): add step to create a pull request with generated prototypes
The commit adds a new step to the workflow file `generate_prototypes_from_TODO.yml`. This step is responsible for creating a pull request with the generated prototypes. The pull request is created using the `peter-evans/create-pull-request` action. The commit message and title for the pull request are set to "Add generated prototypes", and the pull request is created on the `prototypes` branch.