chore(generate_prototypes_from_TODO.yml): update commit and push step to include a meaningful commit message and push changes to the remote repository
The commit and push step in the workflow file has been updated to include a meaningful commit message that indicates the purpose of the commit. The commit message now reads "[Auto] Generate Prototypes for @TODO tasks in /prototypes". This commit message will be used when committing the changes made during the workflow run. Additionally, the changes will be pushed to the remote repository using the "git push" command.